Adde mainmenu - Level - gameover menu sequence
