+++ rollup.config.dist.js
@@ -8,7 +8,7 @@ export default {
 
     //  Our games entry point (edit as required)
     input: [
-        './src/game.ts'
+        './src/app.ts'
     ],
 
     //  Where the build file is to be generated.
@@ -16,7 +16,7 @@ export default {
     //  You can also use 'umd' if you need to ingest your game into another system.
     //  The 'intro' property can be removed if using Phaser 3.21 or above. Keep it for earlier versions.
     output: {
-        file: './dist/game.js',
+        file: './dist/app.js',
         name: 'MyGame',
         format: 'iife',
         sourcemap: false,
