+++ tsconfig.json
@@ -1,12 +1,15 @@
 {
     "compilerOptions": {
-        "target": "es5",
-        "moduleResolution": "node",
-        "sourceMap": true,
-        "inlineSourceMap": true,
-        "inlineSources": true,
+      "noImplicitReturns": true,
+      "noImplicitAny": false,
+      "noEmitOnError": true,
+      "removeComments": false,
+      "sourceMap": true,
+      "target": "es5",
+      "lib": [ "es5", "dom", "es2015.promise", "scripthost" ]
     },
     "include": [
-        "./src/**/*"
+      "lib/**/*",
+      "src/**/*"
     ]
-}
+  }
